+++ services/database/examples_and_test_scripts/database_atp_adw_examples.sh
@@ -47,56 +47,6 @@ SUBNETID="subnetID"
 NSGID="["nsgID1"]"
 UPDATENSGID="["nsgID2","nsgID3"]"
 
-##############################################################################AutonomousDataWarehouse##############################################################################
-
-echo 'Starting Autonomous DataWarehouse Examples'
-
-echo 'Create AutonomousDatawarehouse...'
-CREATE_ADW=$(oci db autonomous-data-warehouse create -c $COMPARTMENT_ID --db-name $DB_NAME1 --admin-password $PASSWORD1 --cpu-core-count $CPU \
-                    --data-storage-size-in-tbs $STORAGE --display-name $DISPLAY_NAME1 --license-model $LICENSE_TYPE \
-                    --wait-for-state AVAILABLE)
-
-ADW_ID=$(jq -r '.data.id' <<< "$CREATE_ADW")
-
-echo "Created AutonomousDatawarehouse with OCID:"
-echo $CREATE_ADW
-
-echo 'Get AutonomousDatawarehouse'
-oci db autonomous-data-warehouse get --autonomous-data-warehouse-id $ADW_ID
-
-echo 'List all AutonomousDatawarehouses in compartment'
-oci db autonomous-data-warehouse list --compartment-id $COMPARTMENT_ID
-
-echo 'List all AutonomousDatawarehouses in compartment in AVAILABLE state'
-oci db autonomous-data-warehouse list --compartment-id $COMPARTMENT_ID --lifecycle-state AVAILABLE
-
-echo 'List 2 AutonomousDatawarehouses in compartment'
-oci db autonomous-data-warehouse list --compartment-id $COMPARTMENT_ID --limit 2
-
-echo 'List all AutonomousDatawarehouses in compartment with specific display name, in descending order'
-oci db autonomous-data-warehouse list --compartment-id $COMPARTMENT_ID --sort-by $DISPLAY_NAME1 --sort-order DESC
-
-echo 'Update AutonomousDatawarehouse DisplayName'
-oci db autonomous-data-warehouse update --autonomous-data-warehouse-id $ADW_ID --display-name $DISPLAY_NAME2
-echo 'Updated AutonomousDatawarehouse DisplayName'
-
-echo 'Update AutonomousDatawarehouse cpuCoreCount and storage'
-oci db autonomous-data-warehouse update --autonomous-data-warehouse-id $ADW_ID --data-storage-size-in-tbs $SCALED_STORAGE \
-                --cpu-core-count $SCALED_CPU --wait-for-state AVAILABLE
-echo 'Updated AutonomousDatawarehouse cpuCoreCount and storageSize'
-
-echo 'Generate and download AutonomousDatawarehouse wallet'
-oci db autonomous-data-warehouse generate-wallet --autonomous-data-warehouse-id $ADW_ID  --password $PASSWORD1 --file  wallet_adw.zip
-
-echo 'Delete AutonomousDatawarehouse'
-oci db autonomous-data-warehouse delete --autonomous-data-warehouse-id $ADW_ID --force --wait-for-state TERMINATED
-echo 'Deleted AutonomousDatawarehouse'
-
-echo 'Trying to Get Deleted AutonomousDatawarehouse. Should not find it.'
-oci db autonomous-data-warehouse get --autonomous-data-warehouse-id $ADW_ID
-
-echo 'End of AutonomousDatawarehouse Examples.'
-
 ##############################################################################Autonomous Transaction Processing##############################################################################
 
 echo 'Starting Autonomous Transaction Processing Examples'